Fix typing

+++ homeassistant/components/overkiz/water_heater/hitachi_dhw.py
@@ -45,29 +45,28 @@ class HitachiDHW(OverkizEntity, WaterHeaterEntity):
     _attr_operation_list = [*OPERATION_MODE_TO_OVERKIZ]
 
     @property
-    def current_temperature(self) -> int | None:
+    def current_temperature(self) -> float | None:
         """Return the current temperature."""
         current_temperature = self.device.states[OverkizState.CORE_DHW_TEMPERATURE]
 
         if current_temperature:
-            return current_temperature.value_as_int
+            return float(current_temperature.value_as_int)
 
         return None
 
     @property
-    def target_temperature(self) -> int | None:
+    def target_temperature(self) -> float | None:
         """Return the temperature we try to reach."""
         target_temperature = self.device.states[
             OverkizState.MODBUS_CONTROL_DHW_SETTING_TEMPERATURE
         ]
         if target_temperature:
-            return target_temperature.value_as_int
+            return float(target_temperature.value_as_int)
 
         return None
 
     async def async_set_temperature(self, **kwargs: Any) -> None:
         """Set new target temperature."""
-
         await self.executor.async_execute_command(
             OverkizCommand.SET_CONTROL_DHW_SETTING_TEMPERATURE,
             int(kwargs[ATTR_TEMPERATURE]),
